Separate out ndex client tests
